+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Upgrade to messaging-lib 0.2.0 for email subject prefix support. (CIS-3351)
+- Use the value of `octri.messaging.default-sender-address` property if `octri.authentication.account-message-email` is not set.
+
 ## [3.0.0] - 2025-09-29
 
 ### Added